=More library work

+++ src/PAA3905.h
@@ -77,7 +77,7 @@ class PAA3905 {
 
         void setResolution(uint8_t res);
 
-        uint8_t getResolution();
+        float getResolution();
 
         void reset();
